requirements.txt
argh==0.26.2
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
configparser==4.0.2
docker-pycreds==0.4.0
docopt==0.6.2
gitdb2==2.0.6
GitPython==3.0.4
gql==0.1.0
graphql-core==2.2.1
idna==2.8
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mosestokenizer==1.0.0
numpy==1.17.2
nvidia-ml-py3==7.352.0
olefile==0.46
openfile==0.0.7
pathtools==0.1.2
Pillow==6.2.0
promise==2.2.1
psutil==5.6.3
pycparser==2.19
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
Rx==1.6.1
sentry-sdk==0.13.0
shortuuid==0.5.0
six==1.12.0
smmap2==2.0.5
subprocess32==3.5.4
toolwrapper==1.0.0
torchvision==0.4.1a0+d94043a
tqdm==4.36.1
urllib3==1.25.6
wandb==0.8.13
watchdog==0.9.0
